+++ src/main/java/amp/awec/command/clipboard/CommandPaste.java
@@ -20,15 +20,11 @@ public class CommandPaste implements CommandManager.CommandRegistry {
 		dispatcher.register(
 			(ArgumentBuilderLiteral) ArgumentBuilderLiteral.literal("/paste")
 				.requires(source -> WorldEditPermissions.canUseWorldEdit((CommandSource) source))
+				.requires(source -> WorldEditPermissions.hasClipboard((CommandSource) source))
 				.executes(context -> {
 					CommandSource source = (CommandSource) context.getSource();
 					CommandPlayerData playerData = CommandPlayerData.get(source, false);
 					if (playerData == null) {
-						return 0;
-					}
-
-					if (playerData.data.clipboardBuffer == null) {
-						MessageHelper.error(source, "Clipboard is empty");
 						return 0;
 					}